Shadowed variables are marked as unused
To account for shadowed variables we have to do better job when merging
results of applying two expressions, because that's the only point where
we can detect duplications arising from applying same env to left and
right parts of "apply" operation.

Old approach does not work because we cannot make list of declared/used
variables unique - because if variable is shadowed it will appear twice
in the list.